Auth_api/dao/IAuthOTPService:update name to IAuthOTPDao, remove: takeResponse method, cut: checkTimeOut method to Auth_api/api/api_confirmOTP/ConfirmOTPService ; Auth_api/ultis add: ServiceUltils with takeResponse method
