+++ src/main/java/com/kaiasia/app/service/Auth_api/dao/IAuthOTPDao.java
@@ -8,10 +8,7 @@ import com.kaiasia.app.service.Auth_api.model.OTP;
 import java.sql.Timestamp;
 import java.util.HashMap;
 
-public interface IAuthOTPService {
+public interface IAuthOTPDao {
     OTP getOTP(HashMap<String, String> body) throws Exception;
     void setConfirmTime( Timestamp now, OTP otp) throws Exception;
-    boolean checkTimeOut(Timestamp endTime);
-    ApiResponse takeRespose(Auth3Response response, ApiError error);
-    ApiResponse takeRespose(Auth3Response response);
 }
